Classes CakePHP App working based on requirments v1
Create a CakePHP app that tracks students taking courses, which teachers are teaching and the total credits for the current semester and the overall total.

This is the first commit which has the all the base logic finished, the only thing that needs to be done is the refinement of the views.

Resolves: #001 Student app request
See also: n/a